Fix init() function signature in the example plugin. Otherwise, it's not being called.
Review URL: http://codereview.chromium.org/3137004

+++ example/example.cc
@@ -166,7 +166,7 @@ class MyInstance : public pp::Instance, public MyFetcherClient {
     }
   }
 
-  virtual bool Init(size_t argc, const char* argn[], const char* argv[]) {
+  virtual bool Init(uint32_t argc, const char* argn[], const char* argv[]) {
     return true;
   }
